FBX, OBJ & DAE - Models & Rigs (#5)
* FBX, OBJ & DAE Models & Rigs

* - Victor_rig_01.FBX - minor fixes in hierarchy and skeleton for blender, unity and assimp (asset import lib)

* - Victor_rig_01.3ds - new 3ds export

* New Directory Structure for File Formats
This was the best possible solution I could come up with (without duplicating tons of files and bloating the repo with multiple copes of each file)
In each directory of assets, where new models/rigs have been made using new file formats, the file formats have been divided by subdirectory
This should allow for easy picking and choosing of what assets to open/look for based upon the program being used.
Yes, there will be some bloat because, for example, if you only need the Maya files, you'll be downloading the rest anyway-
But I have no desire to maintain multiple branches and track changes across all of those branches for other assets, such as the scripts.
If there is a better solution, please feel free to suggest- otherwise, if we merge this into the main branch of cyb3rdog's fork, this will enable
the PR for this fork to be merged into main of the DDL official repo.

* Undo changes to .ma files specifically
Some scripts in the /tools dir reference the directories that store .ma files specifically. I think down the road we will want to update the scripts for further cleanup but, failing that, this is a good start. Undoing changes to the .ma files and their directories for consistency in the scripts.

* Clean up
- wawefront OBJ material (*.mtl) files moved to OBJ sub-dirs
- assets/rigs/Base exports cleaned up

Co-authored-by: Robbie B <65207496+bussardrobbie@users.noreply.github.com>